feat&fix: cal GCD when create price key to get same key when comparePrice(A,B) get 0.
Because snapshort don't have user defined comparator as leveldb, in order to
ensure that snapshort and leveldb have the same put and get operating logic
for the same key, we will calculate GCD when create price key.

In the snapshot, there may be some same keys.
If we use Map to get all the data, the later will overwrite the previous value.
So, if we use list, we need to exclude duplicate keys.